op-build: Move kernel defconfig out of kernel tree
Currently the kernel configuration is in the main kernel tree. This
makes it difficult for third parties to customise the petitboot kernel
configuration as they would need to either change op-build to use a
customised kernel tree or generate patches to the kernel tree to
change the config.

These methods are cumbersome. This patch instead uses some existing
buildroot functionality to move the kernel configuration files into
the openpower overlay configs directory which can be more easily
modified.

Signed-off-by: Alistair Popple <alistair@popple.id.au>