Bump buildroot to 2015.08.1
This moves us to the latest upstream stable release, reducing the number of
local patches we carry, and updating the versions of some packages. The diff
from "op-build external-deps" for a habanero build we get:

  -binutils-2.24.tar.bz2
  +binutils-2.25.1.tar.bz2

  -ethtool-3.18.tar.xz
  -eudev-3.0.tar.gz
  +ethtool-4.0.tar.xz
  +eudev-3.1.2.tar.gz

  ethtool 4.0 release notes:
  Version 4.0 - May 31, 2015

	* Fix: Formatting of RX flow hash indirection table when size not
	  divisible by 8 (-x option)
	* Fix: Add missing Advertised speeds (no option and -s option)
	* Feature: Add support to get expansion ROM version (-i option)
	* Feature: Include SFP serial number and date in EEPROM dump
	  (-m option)

  (over 3.18, from http://git.kernel.org/cgit/network/ethtool/ethtool.git/tree/NEWS )

  -gawk-4.1.2.tar.xz
  -gcc-4.9.2.tar.bz2
  -gettext-0.19.4.tar.xz
  +gawk-4.1.3.tar.xz
  +gcc-4.9.3.tar.bz2
  +gettext-0.19.5.1.tar.xz

  -i2c-tools-3.1.1.tar.bz2
  +i2c-tools-3.1.2.tar.bz2

  +librtas-1.3.13.tar.gz

  -linux-firmware-f66291398181d24856fd2d19454d246199abd5ea.tar.gz
  -lzo-2.08.tar.gz
  +linux-firmware-3161bfa479d5e9ed4f46b57df9bcecbbc4f8eb3c.tar.gz
  +lzo-2.09.tar.gz

  -mpfr-3.1.2.tar.xz
  +mpfr-3.1.3.tar.xz

The binutils jump comes courtesy of upstream e2b99fa1904988e0eeaccd4c98a8e5ec782952f3 which is courtesy of our own Sam Bobroff, and is where we want to go.

This also syncs our openpower/device_table.txt with upstream.

With assistance from Stewart Smith.