packages: Add proper definitions for common-p8-xml and openpower-mrw
The Config.in files for common-p8-xml and openpower-mrw are empty, and
not included by any of the current kconfig infrastructure.

We want to express config dependencies that include these packages, as
we need the xml and mrw components to be build before the PNOR image. In
order to express those dependencies in a DEPENDENCIES variable, we need
to have corresponding kconfig symbols.

For openpower-mrw, we just enable if hostboot is selected.

For common-p8-xml, we enable if hostboot is selected, but also depend on
the MRW package, as this is required by each of the platform xml
packages. This allows those packages to depend on the common package,
rather than open-coding those dependencies in each.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>