Add quiet to skiroot kernel command line
Some BMCs have horrifically slow vuarts (or have enough debug options enabled
so that the vuart is horrifically slow), so slow that a 56k modem would
be an upgrade. Due to the amount of early boot messages that Linux outputs,
this can slow down boot a *lot* - in the order of *over* 20 seconds.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>