openpower: Run firmware whitelist for all machines
Firenze has a script that ensures only the correct firmware is present
in the lib/firmware directory of skiroot. This is useful functionality
for all machines, so rename it and add it to the
ROOTFS_POST_BUILD_SCRIPT config option for all machines.

Note that this means adding new drivers that have loadable firwmare in
the future will require updating this script.

Signed-off-by: Joel Stanley <joel@jms.id.au>