Release Linux 4.9.2-openpower1
This moves from the 4.4 based kernel series to the newer long term
support kernel, Linux 4.9.

To make the move we regenerate the kernel defconfig for the new kernel.

We still have three out of tree patches. This commit makes the following
changes:

  * Drops 5 of our out of tree patches
  * Two remain for XHCI kexec issues
  * One remains for VGA handover

We also update bnx2x firmware version for newer kernel driver, which
means modify the firmware whitelist.

Signed-off-by: Joel Stanley <joel@jms.id.au>