VERSION: Improve logic behind and information in VERSION partition
The VERSION partition is constructed in pkg-versions.mk, and has
the purpose of indicating to both the user and developer/support
what firmware build is on a particular machine.

This patch addresess a few issues with VERSION.

Firstly, the pkg-versions.mk code was making needlessly excessive calls
to git, which could be solved by using "git describe --always --dirty"
rather than going around and around.

Secondly, the information in VERSION was not always useful. Previously,
we'd get an entry like 'open-power-habanero-78d8928', which is only
useful if you posses a tree with that commit in it. With this patch,
instead you get 'open-power-habanero-v1.14-45-g78d89280c3f9' which tells
us a *lot* more information about what was used to build this image. The
same issue was with buildroot (was: 'buildroot-6b9fd80',
now: 'buildroot-2016.11-4-g6b9fd8036d72')

Thirdly, the (truncated) SHA512 hash of the patches to a package was
not always clear that it was a SHA of patches. We now prefix the
patch sha with a 'p' (Git prefixes a sha with a 'g'). For example:
'habanero-xml-6a78496-67a075c' becomes 'habanero-xml-6a78496-p67a075c'

Fourthly, the version targets threw some shell errors, so they're also
fixed.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>