Release Linux 4.10.1-openpower1
This moves all machines to the 4.10 kenrel series.

There are two patch series on top of the stable tree.

The first is the XHCI USB fixes we have carried for sometime.

The second is a 14 patch series that improves the Aspeed VGA driver
including adding support for the ast2500 and configuration of the driver
through the device tree, which is required when the P2A bridge is
disabled.

Signed-off-by: Joel Stanley <joel@jms.id.au>