sbe: Cleanup build commands and pass SBE_COMMIT_ID to the build directly
The sbe build depends on being able to parse the git repo for commit
information. Since the sbe download being used in the build is a git
achive and not a real git repo, this parsing fails. Instead we pass the
commit id directly to bypass the repo requirement.

This change includes a patch containing
open-power/sbe@695eed7bf424f58899e2650ecfdadc5389bf35f5
as updating to the lastest sbe code caused trouble in my testing.

Also, reindent to match other buildroot packages

Signed-off-by: William A. Kennington III <wak@google.com>