Integrate sb-signing-utils into op-build
This adds support for building sb-signing-utils from github.com/hellerda.

It includes the openpower/package/sb-signing-utils/ directory with Config.in,
*.mk, and ./keys directory holding the IBM Imprint keys and README.  Build is
activated by the presence of BR2_OPENPOWER_SECUREBOOT_ENABLED in the _defconfig.
In this patch it is set in habanero_defconfig only.

Even with BR2_OPENPOWER_SECUREBOOT_ENABLED, the package will only be built, it
will not be activated (i.e. used in PNOR packaging step) until additional
patches are added to op-build and openpower-pnor packages.

Signed-off-by: Dave Heller <hellerda@linux.vnet.ibm.com>