Redbud platform: temporary variant of Witherspoon
Witherspoon comes in two variants: 6GPU and 4GPU.

These are referred to as Sequoia and Redbud.

There are minor differences (in how the NVLink bricks are wired up)
that we should be able to handle mostly 'transparently' in firmware.
The current situation is that we need a separate Redbud build that hard
codes some things in skiboot while we wait for Hostboot to populate HDAT
correctly.

Notably, the same image will *NOT* curretly work on Sequoia and Redbud,
which is why we add a redbud platform here.

It is likely that this platform will disappear in a couple of weeks.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>