hostboot: Add patches to fix cmpi assembly
This is change is a noop with the existing toolchain. We change from
implicitly generating the cmpwi to explicitly stating in for
compatibility with newer toolchains.

The patch has been applied to the master branch of hostboot. The P8
backport is pending review.

With gcc 4.9.3, binutils 2.25.2:
```
$ cat asm-test.S
.text
.global test
        cmpi 0, 8, 1
$ /opt/cross/gcc-4.9.3-nolibc/powerpc64-linux/bin/powerpc64-linux-gcc -c asm-test.S
$ objdump -d asm-test.o
0000000000000000 <.text>:
   0:	2c 08 00 01 	cmpwi   r8,1
```

Old toolchain, updated instruction:
```
$ cat asm-test.S
.text

.global test
        cmpwi 0, 8, 1
$ /opt/cross/gcc-4.9.3-nolibc/powerpc64-linux/bin/powerpc64-linux-gcc -c asm-test.S
$ objdump -d asm-test.o
0000000000000000 <.text>:
   0:	2c 08 00 01 	cmpwi   r8,1
```

And then the new toolchain (gcc 6.3.0, binutils 2.28) with the updated asm:
```
$ cat asm-test.S
.text

.global test
        cmpwi 0, 8, 1
$ powerpc64-linux-gnu-gcc -c asm-test.S
$ objdump -d asm-test.o
0000000000000000 <.text>:
   0:	2c 08 00 01 	cmpwi   r8,1
```

Signed-off-by: Joel Stanley <joel@jms.id.au>