Base MEMD changes
This commit adds in the base changes for a new PNOR section.
Currently, the hostboot changes are not complete for the support,
but this commit will enable the section to be filled with blank
data and will unblock testing.  Once hostboot changes are complete,
there will need to be a change in the pnor project.