Integrate sb-signframework into op-build
This adds support for building sb-signframework from github.com/cjengel.

It includes the openpower/package/signframework/ directory with Config.in and
*.mk,  Build is activated by the presence of BR2_OPENPOWER_SECUREBOOT_ENABLED
in the _defconfig. plus BR2_OPENPOWER_SECUREBOOT_SIGN_MODE=production.  In
this patch it is set in habanero_defconfig only, left commented out.

Even with these configuration properties set, the package will only be built,
it will not be activated (i.e. used in PNOR packaging step) until additional
patches are added to op-build and openpower-pnor packages.

This is a test patch only; I have some questions which I will elaborate in
the Pull Request discussion.

Signed-off-by: Dave Heller <hellerda@linux.vnet.ibm.com>