openpower-pnor: build a PNOR squashfs as consumed by OpenBMC
Recently, OpenBMC has started supporting a firmware update mechanism
that requires a specially formatted tarball to be passed to it.

Since we want op-build to produce images that are easily consumable,
this patch makes the openpower-pnor build produce one of these images.

There is currently no documentation of this format, although the code
that produces it from a PNOR is open source.

Signed-off-by: Charles P. Hofer <charles.hofer@ibm.com>
[stewart@linux.vnet.ibm.com: flattened commit history, rewrote commit message]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>