Test PR to make sure new attr/targ xml generation works w/ OP
Hostboot changed out its attribute and targeting xml is genearted.
The xml that hostboot makes gets exported to the op-build process.
There have been some changes to allow op-build to handle the new
format of the attr/targ xml it gets from hostboot. This commit
tests that this path works. Do not merge.