kernel: Move to Linux v4.14.1-openpower1
Integrate v4.14.1 stable release

Update P9 kernel defconfig with 4.14 related options:

 Harden slab freelist metadata (SLAB_FREELIST_HARDENED)=y
 IBM Virtual Accelerator Switchboard (VAS) (PPC_VAS)=n
 Network Service Header (NSH) protocol (NET_NSH)=n
 Mellanox Technologies MLX5 MPFS support (MLX5_MPFS)=n
 Mellanox Technologies MLX5 SRIOV E-Switch support (MLX5_ESWITCH)=n
 RmNet MAP driver (RMNET)=n
 Large payload keys (BIG_KEYS)=n

We build in the MLX5 driver but the help text suggests we do not need
those features compiled in for petitboot.

We do not use the VAS features in petitboot.

I40E is required by Boston and P8 SMC machines, so it is retained in the
configuration.

IGB is dropped as it was not enabled when moving to 4.14 and no one
noticed.

The P8 kernel defconfig is updated to be in line with the P9 defconfig.
The only difference at this stage is the default commandline which
contains workarounds for P9:

 -CONFIG_CMDLINE="console=tty0 console=hvc0 powersave=off"
 +CONFIG_CMDLINE="console=tty0 console=hvc0 quiet"

Signed-off-by: Joel Stanley <joel@jms.id.au>