gcc: disable libsanitizer
GCC builds this library in order to enable the -fsanitize error detection
features. This part of the compile takes several minutes on a Westmere
Xeon build server, and to make it worse the build serialises at that
point.

Our GCC is 6.4 at this point. Running 'op-build toolchain' shows a
saving of about 33% (4 minutes) on a 40 core machine and a 25% (6
minutes) on a 2 core machine.

We build neither the kernel nor userspace with any of these tools, and
probably never will in production.

In order to save build time disable libsantizer for GCC builds.

Signed-off-by: Joel Stanley <joel@jms.id.au>