hostboot: Drop patch that is applied upstream
hostboot-0013-htmgt-Remove-unused-variable has been applied upstream as
of hostboot commit 72aa4a65f70deeab50390b5b3040cc04d24cde98 so we don't
need to patch it into the tree.

The upstream version had a different Gerrit commit ID, so it did not
automatically get dropped as it should have.

Note that a failing patch causes subsequent patches to not apply (see
https://github.com/open-power/op-build/issues/1194), but luckily this
was the last one.

Signed-off-by: Joel Stanley <joel@jms.id.au>