Optimize Busybox configuration
There's a number of things we do *not* need in the limited
petitboot environment, so we should exclude them.

We also enable the taskset utility as we can use it in op-test
to help checking cpu stop states, and that's reason enough.

Before 989872 (358760 xz compressed)
After  918816 (339628 xz compressed)

A saving of 19132 bytes!

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>