kernel: Move to Linux 4.16.1-openpower2
Move to the 4.16 series. Changes to patches:

 - "tg3: APE heartbeat changes" patch was merged upstream and can be
 dropped

 - "powerpc/mm/radix: Fix checkstops caused by invalid tlbiel" was found
 when testing with 4.16.1-openpower1 and has been posted upstream

A few new options are turned off by default as they are not useful to
petitboot. In the future we may enable DEFERRED_STRUCT_PAGE_INIT, but it
defaults to off for now.

We bump the buildroot repository in order to support the new kernel
version.

Signed-off-by: Joel Stanley <joel@jms.id.au>