hostboot: Build with OP920 "stable" branch by default, master as on option
This promotes our efforts to stabilise op-build towards the v2.0
release.

We have a config option to select between living on the edge and the
stable tree.

In the future, we expect this to move as the development model evolves.

To build with Hostboot as 'master', BR2_HOSTBOOT_LATEST_VERSION=y
is set.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>