linux: Enable ipmi watchdog support
We need this support so that the skiroot linux kernel can handle the
watchdog after it is handed off by skiboot. This way we can keep the
watchdog enabled throughout the entire skiboot handoff to linux instead
of disabling it prior to executing the linux kernel. This allows us to
increase our watchdog coverage during boot.

It also gives us the benefit that we can now use the ipmi watchdog
through /dev/watchdog in the skiroot environment.