Enable BR2_LINUX_KERNEL_XZ=y for all platforms
We keep creeping *just* over the size for BOOTKERNEL on our P8
builds due to some non-determinism in the kernel and initramfs
build process.

It turns out we didn't have the *kernel* being built with XZ
compression (just the initramfs), so if we enable that, we get
close to a megabyte back, which gives us some breathing room.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>