hostboot: enable use of ccache
We were missing the CCACHE part for the HB cross compiler, which
was meaning CCACHE wasn't getting invoked (if enabled) for building
hostboot.

With my "deterministic xml parsing" patch for Hostboot (not yet
upstream), this takes the time for re-compiling hostboot from 15 minutes
to less than 3 minutes (at least on this machine).

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>