buildroot: Move to 2018.08.2
Upstream changelog:

 https://git.buildroot.net/buildroot/plain/CHANGES?id=2018.08.2

Set BR2_GCC_VERSION_6_X=y in all of the configs that include hostboot,
as the default has moved to GCC 7 and hostboot cannot build with a
modern compiler.

We are carrying a patch to dtc in our buildroot tree that fixes a
segfault when converting the powernv device tree. It has been submitted
upstream to buildroot (and the patch itself is upstream in dtc, but not
a released version).

Host packages updated:

  libxml2: 2.9.7 -> 2.9.8
  lz4: v1.8.1.2 -> v1.8.2
  zstd: v1.3.3 -> v1.3.5
  dtc: 1.4.4 -> 1.4.7

Shipped packages updated:

  busybox: 1.29.0 -> 1.29.2
  linux-firmware: 65b1c68c63f974d72610db38dfae49861117cae2 -> 8d69bab7a3da1913113ea98cefb73d5fa6988286
  ethtool: 4.15 -> 4.16
  util-linux: 2.32 -> 2.32.1
  dtc: 1.4.4 -> 1.4.7
  elfutils: 0.169 -> 0.171

Signed-off-by: Joel Stanley <joel@jms.id.au>
(cherry picked from commit 19ab137a9b16e6884605cb4faec445f5ded76d5b)
Signed-off-by: Joel Stanley <joel@jms.id.au>