ci: Add release mode to build legal info
Add an option to build the Buildroot legal-info target after building a
platform. This is mainly useful as part of building platforms for
inclusion in images for distribution, but is also good for CI systems
doing a sanity check on what is being built.
This also updates build-all-defconfigs.sh to use getopt to handle its
inputs more sanely.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>