openpower: Default PNOR_SQUASHFS to n
Witherspoon is the only system that uses this option. Instead of
updating all of the non-Witherspoon systems to turn it off, this commit
defaults it to 'n' and updates the witherspoon defconfigs to enable it.

The change to zz is due to the new default.

Signed-off-by: Joel Stanley <joel@jms.id.au>
(cherry picked from commit b9859a023e96b7917a9ca6e42dac82bbd7f0aec0)
Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>