ci: Get kernel version
When building with the SDK the build scripts have to specify the
kernel version (and GCC). This was hardcoded, which is fine as the CI
scripts are checked into the repo where the changes are made.

Hoover it's one more thing to remember when bumping the kernel version.
This grabs it from the BR2_LINUX_KERNEL_CUSTOM_VERSION_VALUE in the
given config file, which should always be set.

Signed-off-by: Joel Stanley <joel@jms.id.au>
(cherry picked from commit a81c124cb68903e5b14d981d0cd84c0581fe92f8)
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>