Pull in P8 PNOR XML Layout Fix for HBD/HBD_RW Partitions
This commit pulls in an update to the P8 PNOR XML Layout files
that were causing build problems for the VESNIN configuration.  The
change is that the combined size of the HBD and HBD_RW partitions
was redistributed to HBD and HBD_RW to allow for the fact that for
P8 systems there us much more read-only attribute/targeting content
then read-write.

Signed-off-by: Mike Baiocchi <mbaiocch@us.ibm.com>