Add op-test-framework as a git-submodule
The idea is we tie an op-build commit to a op-test commit so that we
control what tests people run in CI via code in op-build repo rather
than random jenkins jobs.

This is a starting point to get to the ponit where we have an op-test
commit alongside an op-build commit, where the op-build commit is meant
to pass the op-test tests.

We start with op-test-framework v0.13

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>