Add op-build script
The current build instructions rely on sourcing op-build-env into the
user's shell. But that assumes the user's shell is bash or sh. There's
also no reason to pollute the user's shell with op-build related
environment variables etc.

So add an op-build script, which performs the same function and
doesn't depend on what the user's shell is and is self contained. It
just uses op-build-env, which is unchanged, so should have no impact
on CI or other existing workflows.

Update the build instructions to refer to it.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>