linux: Re-enable STRICT_KERNEL_RWX
A workaround was put in place in future kernels (as of 5.0.10) for the
issue exposed by STRICT_KERNEL_RWX. The workaround changes RADIX to
require HUGETLBFS while the root cause is fixed.

The disabling of STRICT_KERNEL_RWX can be reverted with this alternate
workaround in place.

We need to enables HUGETLBFS so we continue using the radix MMU,
otherwise the configuration will fall back to using hash MMU.

See https://github.com/linuxppc/issues/issues/237.

Fixes: 885ef25a0d68 ("linux: temporarily disable STRICT_KERNEL_RWX")
Fixes: 85cc33affffd ("kernel: Move to Linux v4.14.1-openpower1")
Signed-off-by: Joel Stanley <joel@jms.id.au>