linux: Re-enable radix MMU
A workaround was put in place for the issue exposed by STRICT_KERNEL_RWX
which meant RADIX required HUGETLBFS. As we didn't have HUGETLBFS
enabled the petitkernel defaulted to radix.

This enables HUGETLBFS so we go back to using the radix MMU.

The disabling of STRICT_KERNEL_RWX can be reverted with this alternate
workaround in place.

See https://github.com/linuxppc/issues/issues/237.

Fixes: 885ef25a0d68 ("linux: temporarily disable STRICT_KERNEL_RWX")
Signed-off-by: Joel Stanley <joel@jms.id.au>