skiboot: Fix building device tree with host-dtc
We need to do the build with the PATH set to the host tools. Without
this change the build fails when the host lacks a dtc, as it is not in
the PATH.

I suspect CI missed this as only opal_defconfig tries to build with the
SKIBOOT_DEVICETREE option and the CI that builds this config also
happened to have dtc installed:

$ git grep SKIBOOT_DEVICETREE openpower/configs/
openpower/configs/opal_defconfig:BR2_SKIBOOT_DEVICETREE=y

Note that this will still fail without a skiboot that has "devicetree:
Don't set path to dtc in makefile"[1] applied. This patch won't make any
currently working system fail, so it is safe to merge while we wait for
the skiboot patch to be merged.

[1] https://patchwork.ozlabs.org/patch/1097826/

Signed-off-by: Joel Stanley <joel@jms.id.au>