Move to Linux v5.1.15-openpower1 and buildroot 2019.02.3
This updates to a in-support stable Linux release, resolving potential
security and stability issues.

Buildroot stays on the same major version with the .2 and .3 stable
releases added in.

The skiroot defconfig is updated to ensure we still run the MMU in Radix
mode (see http://git.kernel.org/torvalds/c/8adddf349fda0). It also
disables xmon by default.

Signed-off-by: Joel Stanley <joel@jms.id.au>