Un-ship iprutils
Petitboot has pulled in iprutils for the iprconfig tool but there is no
code in Petitboot that uses it. Furthermore, openpower systems do not
have this hardware so there is no need for it to be included.

Add it to the FSP based zz and firenze configurations as it may be used
there.

Signed-off-by: Joel Stanley <joel@jms.id.au>