Ship PCI Utils
This adds the PCIUTILS package back to the petitroot. In the past it was
included in the build as a dependency of IPRUTILS.

When that was dropped from the build we inadvertently dropped PCIUTILS,
which contains lspci. We have an lspci in busybox but as Oliver pointed
this out this is not adequate:

 > why did we replace the useful lspci in skiroot with the useless lspci?
 > the busybox lspci doesn't even show the domain number

Fixes: a1ae0bc353b8 ("Un-ship iprutils")
Signed-off-by: Joel Stanley <joel@jms.id.au>