Update openpower-pnor makefile to use BR2_ version vars
Previously we were using the * wildcard to lookup files to copy
over for debug data. This breaks when there are multiple versions
of a given package built in the output/build directory. This commit
will make the openpower-pnor.mk files use the build root env vars
(BR2_..) to make sure we just attempt to copy files from the current
version we are using.

Signed-off-by: Christian Geddes <crgeddes@us.ibm.com>