vesnin: Fix flash@0 reg property in device tree
Since the "ibm,opal" node doesn`t refer to any "bus", the
"#address-cells" and "#size-cells" being zero is correct. However, in
this case, the “reg” property that determines the size of the flash
memory in the child node “flash@0” has an invalid format. For this
reason, the FWTS on the Linux host displays the following warning:

  /ibm,opal/flash@0:reg: property has invalid length (8 bytes)
  (#address-cells == 0, #size-cells == 0)

This skiboot patch sets the #address-cells/#size-cells to 1 and fixes
this problem temporarily. However, in the future, need to add some
property that contains the size of the flash device instead of "reg".

Resolve:
[*] https://github.com/open-power/skiboot/issues/47
[*] https://j.yadro.com/browse/SRV-553
[*] https://j.yadro.com/browse/SRV-632