nicole: Set correct type for planar VPD EEPROM
Nicole platform has 32KiB EEPROM (planar VPD), connected to Proc 0,
Engine 2, Port 0, Address 0x50.

To read/write the VPD on a host, it is required to have the correct
"compatible" property in the Linux kernel device tree.
This property is populated by Skiboot and its value is based on the
I2C device type from HDAT structures of Hostboot.

The current Hostboot implementation sets the type for all EEPROMs to
HDAT_I2C_DEVICE_TYPE_SEEPROM regardless of the I2C_TYPE attribute in
a machine xml.
As a result, the Linux kernel loads the "atmel,24c128" driver for
these EEPROMs, which limits their size to 16KiB.

This patch forces Hostboot to set the I2C type of planar VPD's EEPROM
to "atmel,24c256" (32KiB).

Signed-off-by: Artem Senichev <a.senichev@yadro.com>