[RFC] Backport several Petitboot fixes
As of Jan 2020, there are a lot of Petitboot fixes that were merged
upstream, but not yet picked by Petitboot 1.7.5.

There is an attempt to backport them and hopefully release a Petitboot
1.7.6 sometime soon, but in order to properly test them, I'm applying
the exact same set to op-build-2.0.y (the only user I'm aware of) as a
set of patches first.

If validation succeeds, I'll propose this same set to Petitboot 1.7.x
and propose a different patch to op-build-2.0.y to bump the Petitboot
stable version.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>