openpower-pnor: Add config name to manifest
The config name is generally speaking the name of the system. This
name will be added to the manifest file in the squashfs tarball to
check against the name of the system, in order to log an error if
the machine names don't match, since this might indicate that the
wrong image file is being used for that system.

Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>