Witherspoon platform: Enable Radeon at boot device
There have been requests to enable Skiroot/Petitboot console output
through Radeon cards (specifically Edwards2) on Witherspoon platforms.
So far this was not possible due to bugs in the kernel that made
Edwards2 not kexec-safe.

Those bugs are not fixed, and we can now enable this for the Witherspoon
platform.

Since this patch requires changes to the skiroot kernel config that are
not upstream, and also changes to the PNOR Layout (for a larger
BOOTKERNEL), this change is only applied to the Witherspoon platforms,
leaving all other platforms as they were.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>