Introduce Alternate Toolchain gcc and binutils
In the ideal buildroot world, there is one definition of a Toolchain
that will be used to build the entire set of packages used in a
distribution.

However, in the OpenPower world, this rule is not completely true, due
to:

 1) The need to build for two different target architectures at once
    (ppc64le and ppe42) - this is already addressed by the ppe42-gcc
    and ppe42-binutils packages

 2) The fact that some of the legacy packages are not actively
    maintained, but still needed (such as Hostboot-P8), while other
    packages are actively updated and could benefit using newer
    compilers (Linux Kernel, Skiboot etc).

Considering that (1) is already solved, and that (2) is desirable, this
patch reuses the mechanism used by (1) but to create an "Alternate"
Toolchain (initially composed of gcc and binutils) that could be used by
packages that are not willing or able to move towards newer compilers.

Once in use by a subset of packages, this also may allow for downstream
branches to keep using an older toolchain, while buildroot is re-based,
without necessarily moving the snapshot where those packages were
tested. This could be useful, for example, for upgrading Skiroot
environment completely without touching Hostboot, OCC, HCODE or even
Skiboot for that matter.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>