ci/build-all-defconfigs.sh: Fix output directory handling
There's a longstanding issue in the ci/build-all-defconfigs.sh where it
erroneously uses "or" as part of the test syntax. The command fails and
since it's a sanity check, it doesn't have any consequences, e.g.:

  + for p in ${PLATFORM_LIST}
  + DEFCONFIGS+=($p$CONFIGTAG)
  + '[' -z /output-ubuntu1804 or '!' -d /output-ubuntu1804 ']'
  ./ci/build-all-defconfigs.sh: line 59: [: too many arguments

Even if the original intention seemed to be to check for a non-empty and
valid directory, the script was creating a completely new directory, and
that is apparently the CI's environment assumption here as well.

Remove check altogether since getopt "o:" will always give a non-empty
$OUTDIR, and that directory is expected to be created anyway.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>