nicole: Add testing support for standard PCIe hotplug
An OS may re-enumerate the PCIe tree during boot and after hot-plugging a
bridge, but this leads to de-sync of the cached and actual BDF addresses.

Track the changes in bus numbering and update the cached topology (a tree
consisting of struct pci_device nodes) correspondingly, including PEs.

Append the cached topology when reading a config space reveals a hot-added
device, with no need for an explicit OPAL call for that.

Tested on the Nicole platform with Linux kernel 5.7.0-rc3 patched for using
the standard pciehp hotplug driver and sysfs rescan instead of the pnv_php
driver.

Signed-off-by: Sergey Miroshnichenko <s.miroshnichenko@yadro.com>