Fix Hostboot binutils usage
Hostboot needs to link against a static libfd.a and libiberty.a, but by
default buildroot will build the toolchain with shared libaries only, so
hostboot{-p8} hacks it's way through it by using the static libraries
from the binutils build directory, thus also depending on it's build.

In a scenario where we want to have a re-usable toolchain, any package
depending on host-binutils will force it to be rebuilt (even if we are
providing a perfectly fine binutils as part of our sdk).

Luckly, Hostboot use for those libraries is to build debugging and
tracing code using the HOST compiler (which is not a cross-compiler),
and since both Fedora and Ubuntu provides those packages, we can just
use them instead.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>