Support incremental builds in Hostboot
Previously, on every build, the makefile for the Hostboot package was
updating the timestamp of a Hostboot source file on which nearly all
other Hostboot files have a dependency.  This causes files that
haven't changed to be rebuilt, which takes a lot of time.

This commit changes the makefile to only update the relevant file if
it needs to do so. This speeds up Hostboot rebuilds by a large factor.