p10: Remove CAPP
The CAPP partition was used on P8 and P9 to store the CAPP microcode
that skiboot loads into the CAPP hardware. The CAPP no longer exists
on P10. In skiboot, it's already removed any attempt to load the CAPP
microcode on P10, so any references and dependencies to it can be
removed now.

Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>