RH8: Update the call to python to be python2 in migration effort to Red Hat 8
RH7 has /bin/python defined in it's environment whereas RH8 will not.  RH8
will not have a default of /bin/python but will only have /bin/python2 and
/bin/python3 defined.

Considering that for RH7 /bin/python is just a symbolic link to /bin/python2,
making this change to explicitly call python2 will have no detrimental affects
for compiling SBE on RH7 but will position the compiling of the SBE code on
RH8 for success.

Signed-off-by: Roland Veloz <rveloz@us.ibm.com>