Set The Secure Version to 2 for P10 eBMC Builds
This commit sets the variable BR2_OPENPOWER_SIGNED_SECURITY_VERSION
in the p10ebmc_defconfig configuration to "2" from "0".  This new
value of 2 will be built into the secure headers of the lids.

Once a driver with this Secure Version of 2 is installed on a system,
a user will not be able to boot a driver with a Secure Version less
than 2 with security enabled.

Signed-off-by: Mike Baiocchi <baiocchi@ibm.com>