buildroot: Stable release and fix lvm2 download site
Update to the final 2021.02.12 release to incorporate security and bug fixes
that were made. Note that 2021.02 was declared end of life on 6th April
2022.

In addition, it seems Redhat has taken sources.redhat.com off the
internet. Backport a patch to update the URL to sourceware.org so we can
still build.

buildroot change 6cc35cb59908 ("package/lvm2: default standard install
to y") in 2021.02.4 changed BR2_PACKAGE_LVM2_STANDARD_INSTALL to
enabled by default, which adds a large number of lvm userspace tools.
Disable it in the maintained defconfigs.

Signed-off-by: Joel Stanley <joel@jms.id.au>